backend/pyproject.toml
[project]
name = "api-app-credit"
version = "0.1.0"
description = "A FastAPI application for credit management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]
